Move tasks from eu.xenit.gradle.tasks to eu.xenit.gradle.docker.tasks & eu.xenit.gradle.alfresco.tasks
