+++ src/integrationTest/examples/docker-compose-example/subpr1/build.gradle
@@ -1,7 +1,7 @@
 plugins {
     id "eu.xenit.docker"
 }
-import eu.xenit.gradle.tasks.DockerfileWithCopyTask
+import eu.xenit.gradle.docker.tasks.DockerfileWithCopyTask
 
 task createDockerFile(type: DockerfileWithCopyTask) {
     from('alpine:edge')
