+++ src/integrationTest/examples/docker-compose-example/subpr2/build.gradle
@@ -2,7 +2,7 @@ plugins {
     id "eu.xenit.docker"
 }
 
-import eu.xenit.gradle.tasks.DockerfileWithCopyTask
+import eu.xenit.gradle.docker.tasks.DockerfileWithCopyTask
 
 task createDockerFile(type: DockerfileWithCopyTask) {
     from('alpine:edge')
