+++ src/main/java/eu/xenit/gradle/alfresco/DockerAlfrescoPlugin.java
@@ -2,13 +2,13 @@ package eu.xenit.gradle.alfresco;
 
 import eu.xenit.gradle.docker.DockerBuildBehavior;
 import eu.xenit.gradle.docker.DockerConfigPlugin;
-import eu.xenit.gradle.tasks.DockerfileWithWarsTask;
-import eu.xenit.gradle.tasks.InjectFilesInWarTask;
-import eu.xenit.gradle.tasks.InstallAmpsInWarTask;
-import eu.xenit.gradle.tasks.MergeWarsTask;
-import eu.xenit.gradle.tasks.StripAlfrescoWarTask;
-import eu.xenit.gradle.tasks.WarEnrichmentTask;
-import eu.xenit.gradle.tasks.WarLabelOutputTask;
+import eu.xenit.gradle.alfresco.tasks.DockerfileWithWarsTask;
+import eu.xenit.gradle.alfresco.tasks.InjectFilesInWarTask;
+import eu.xenit.gradle.alfresco.tasks.InstallAmpsInWarTask;
+import eu.xenit.gradle.alfresco.tasks.MergeWarsTask;
+import eu.xenit.gradle.alfresco.tasks.StripAlfrescoWarTask;
+import eu.xenit.gradle.alfresco.tasks.WarEnrichmentTask;
+import eu.xenit.gradle.alfresco.tasks.WarLabelOutputTask;
 import java.util.ArrayList;
 import java.util.List;
 import org.alfresco.repo.module.tool.WarHelperImpl;
