+++ src/main/java/eu/xenit/gradle/alfresco/tasks/InjectFilesInWarTask.java
@@ -1,4 +1,4 @@
-package eu.xenit.gradle.tasks;
+package eu.xenit.gradle.alfresco.tasks;
 
 import de.schlichtherle.truezip.file.TFile;
 import java.io.File;
