+++ src/main/java/eu/xenit/gradle/alfresco/tasks/MergeWarsTask.java
@@ -1,4 +1,4 @@
-package eu.xenit.gradle.tasks;
+package eu.xenit.gradle.alfresco.tasks;
 
 import java.io.File;
 import java.util.Map;
