+++ src/main/java/eu/xenit/gradle/alfresco/tasks/StripAlfrescoWarTask.java
@@ -1,4 +1,4 @@
-package eu.xenit.gradle.tasks;
+package eu.xenit.gradle.alfresco.tasks;
 
 import static eu.xenit.gradle.alfresco.DockerAlfrescoPlugin.LABEL_PREFIX;
