+++ src/main/java/eu/xenit/gradle/alfresco/tasks/WarEnrichmentTask.java
@@ -1,4 +1,4 @@
-package eu.xenit.gradle.tasks;
+package eu.xenit.gradle.alfresco.tasks;
 
 /**
  * Created by thijs on 2/27/17.
