+++ src/main/java/eu/xenit/gradle/alfresco/tasks/WarLabelOutputTask.java
@@ -1,4 +1,5 @@
-package eu.xenit.gradle.tasks;
+package eu.xenit.gradle.alfresco.tasks;
 
 public interface WarLabelOutputTask extends WarOutputTask, LabelSupplierTask {
+
 }
